refactor: move stream handling to dedicated module
