+++ crates/tauri-plugin-mihomo/src/stream.rs
@@ -3,16 +3,85 @@ use std::{
     task::{Context, Poll},
 };
 
+use futures_util::{
+    SinkExt, StreamExt,
+    stream::{SplitSink, SplitStream},
+};
 use pin_project::pin_project;
-use tokio::io::{AsyncRead, AsyncWrite};
 #[cfg(unix)]
 use tokio::net::UnixStream;
 #[cfg(windows)]
 use tokio::net::windows::named_pipe::{ClientOptions, NamedPipeClient};
+use tokio::{
+    io::{AsyncRead, AsyncWrite},
+    net::TcpStream,
+};
+use tokio_tungstenite::{MaybeTlsStream, WebSocketStream, tungstenite::Message};
 #[cfg(windows)]
 use windows_sys::Win32::Foundation::ERROR_PIPE_BUSY;
 
 use crate::{Error, Result};
+
+pub enum WsStream {
+    Tcp(WebSocketStream<MaybeTlsStream<TcpStream>>),
+    Socket(WebSocketStream<SocketStreamKind>),
+}
+
+impl From<WebSocketStream<MaybeTlsStream<TcpStream>>> for WsStream {
+    fn from(value: WebSocketStream<MaybeTlsStream<TcpStream>>) -> Self {
+        WsStream::Tcp(value)
+    }
+}
+
+impl From<WebSocketStream<SocketStreamKind>> for WsStream {
+    fn from(value: WebSocketStream<SocketStreamKind>) -> Self {
+        WsStream::Socket(value)
+    }
+}
+
+impl WsStream {
+    pub fn split(self) -> (WsWriteKind, WsReadeKind) {
+        match self {
+            Self::Tcp(stream) => {
+                let (write, read) = stream.split();
+                (WsWriteKind::Tcp(write), WsReadeKind::Tcp(read))
+            }
+            Self::Socket(stream) => {
+                let (write, read) = stream.split();
+                (WsWriteKind::Socket(write), WsReadeKind::Socket(read))
+            }
+        }
+    }
+}
+
+pub enum WsReadeKind {
+    Tcp(SplitStream<WebSocketStream<MaybeTlsStream<TcpStream>>>),
+    Socket(SplitStream<WebSocketStream<SocketStreamKind>>),
+}
+
+impl WsReadeKind {
+    pub async fn next(&mut self) -> Option<crate::Result<Message>> {
+        match self {
+            Self::Tcp(read) => read.next().await.map(|v| v.map_err(Into::into)),
+            Self::Socket(read) => read.next().await.map(|v| v.map_err(Into::into)),
+        }
+    }
+}
+
+pub enum WsWriteKind {
+    Tcp(SplitSink<WebSocketStream<MaybeTlsStream<TcpStream>>, Message>),
+    Socket(SplitSink<WebSocketStream<SocketStreamKind>, Message>),
+}
+
+impl WsWriteKind {
+    pub async fn send(&mut self, message: Message) -> crate::Result<()> {
+        match self {
+            Self::Tcp(write) => write.send(message).await?,
+            Self::Socket(write) => write.send(message).await?,
+        }
+        Ok(())
+    }
+}
 
 #[pin_project(project = WrapStreamProj)]
 pub enum SocketStreamKind {
